Separate out getPairs selector
