+++ package.json
@@ -252,7 +252,8 @@
       "<rootDir>[/\\\\](www|docs|node_modules|scripts|test)[/\\\\]",
       "<rootDir>/src/utils/network-exporters",
       "<rootDir>/src/utils/networkQuery",
-      "<rootDir>/integration-tests"
+      "<rootDir>/integration-tests",
+      "<rootDir>/platforms"
     ],
     "testEnvironment": "node",
     "testURL": "http://localhost",
